pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "inferrvc"
authors = [{name = "Charles Marks", email = "charlesmarksco@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
requires-python = ">=3.0" #just guessing here.
dynamic = ["version", "description"]
dependencies = [
    'python-dotenv',
    'numpy',
    'scipy',
    'soundfile',
    'torch',
    'torchaudio',
    'librosa',
    'pyworld',
    'praat-parselmouth',
    'torchcrepe',
    'huggingface-hub',
    'fairseq',
    #'https://github.com/One-sixth/fairseq/archive/main.zip; python_version >= "3.11"',
    'faiss-cpu'
]

[project.urls]
Home = "https://github.com/CircuitCM/RVC-inference"

[tool.flit.module]
name = "inferrvc"

